Changes to +context, boosters and custom language flags (#183)
* Update context command

Removed +context view and replaced it with link sheet and list. Added +context show as an alias for +context get. Changed the +context link timeout to 1 minute and added ms as a field

* Updated boosters

* Delete the +context link command usage

* Changed boosters and how flags work

Added a folder for custom language flags and made the flag URL redirect to there. Also changed the link for biscuit's custom languages to include %code% like all the others.
Doubled the chance for booster staff members to show up and added missing boosters

* Changed the remaining occurences of view

* Changed the URL to dot